requirements.txt
asgiref==3.5.2
autopep8==1.7.0
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1
django-axes==5.39.0
django-ckeditor==6.5.0
django-ipware==4.0.2
django-js-asset==2.0.0
filelock==3.8.0
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pytube==12.1.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
virtualenv==20.16.4
whitenoise==6.2.0
